backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.3
email-validator==2.1.1
bcrypt==4.1.3
httpx==0.27.0
pytest==8.2.1
black==24.4.2
ruff==0.5.0
Pillow==10.3.0
python-multipart==0.0.9
aiofiles==23.2.1
